+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+2.4.1
+-----
+- Added compatibility method to config class
+- Added sys_get_temp_dir fallback
+
 2.4.0
 -----
 - Major config system refactoring